A .py version for SB response
This version is the first solid .py file that can return a list of prediction to the Spring Boot server through the Flask file 'flaskIntro.py'. 
This version does not keep any of the training or dataset preparation because it loads the already trained .joblib files and uses its values. 
The limitation to the file is that it returns a list of Strings, which are only numbers and doesn't mean anything to the recipient just yet. 
Formatting the file so that it returns dictionary form should be better.